Extract generateFormPdfBuffer into lib/forms/generate.js
- Add server/lib/forms/generate.js with generateFormPdfBuffer() function
- Move cert narcotics notice appending into FormCoR.generatePdf()
- Import FormNarcoticsNotice from source JSX instead of dist
- Convert dynamic pdf.js and FormNarcoticsNotice imports to static imports in FormCoR.jsx
- Update generateForms.js and routes/api/forms/index.js to use generateFormPdfBuffer()
- Remove pdf-lib and PDFDocument imports from both callers
